career/requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
django-filter==23.1
psycopg2-binary==2.9
cloudinary==1.30.0
python-dotenv==1.0.0